Move MapParam to its own file
