+++ app/src/editor/canvas/components/Ports/MapParam/mapParam.pcss
@@ -1,0 +1,2 @@
+.root {
+}
